requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.0.10
djangorestframework==3.13.1
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
rpds-py==0.13.2
sqlparse==0.4.4
uritemplate==4.1.1
